feat: sub-pixel font size granularity (0.5px steps)
Change font size control from integer to 0.5px increments for finer
granularity on mobile. Update range input step/min/max, switch parseInt
to parseFloat for font size reads, and format display labels to show
one decimal only when fractional.

Closes #31

+++ src/modules/settings.ts
@@ -168,7 +168,7 @@ export function initSettingsPanel(): void {
   }
 
   document.getElementById('fontSize')!.addEventListener('input', (e) => {
-    _applyFontSize(parseInt((e.target as HTMLInputElement).value));
+    _applyFontSize(parseFloat((e.target as HTMLInputElement).value));
   });
 
   const themeSelect = document.getElementById('termThemeSelect') as HTMLSelectElement;
